+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### New features
 * Added data label overlap handling setting
 * Added per-stream color settings
-* Added Label Orientation option to control forced label rotationng
+* Added Label Orientation option to control forced label rotation
 
 ### Fixes
 * Bound data labels to streams